check user to be root to launch dockers

+++ db/stop.sh
@@ -3,6 +3,11 @@
 websrv_container_name="delssajri_websrv"
 couchdb_container_name="delssajri_couchdb"
 
+if [ "$(id -u)" != "0" ]; then
+   echo "This script must be run as root" 1>&2
+   exit 1
+fi
+
 echo "Stopping ${couchdb_container_name}..."
 docker kill $couchdb_container_name
